[D13P2] Part 2: 69694090468468 is too low
That's not the right answer; your answer is too low. If you're stuck, make sure you're using the full input data; there are also some general tips on the about page, or you can ask for hints on the subreddit. Please wait one minute before trying again. [Return to Day 13]

+++ src/Day13.kt
@@ -56,7 +56,7 @@ fun main() {
 
             if ((pX == a * dXA.toBigInteger() + b * dXB.toBigInteger()) && (pY == a * dYA.toBigInteger() + b * dYB.toBigInteger())) {
                 println()
-                cost += (a * BigInteger("2") + b).toInt()
+                cost += (a * BigInteger("2") + b).toLong()
             } else {
                 println(" -xX not a solution")
             }
